requirements.txt
numpy==1.24.3
matplotlib==3.7.1
pillow==10.0.0
tqdm==4.65.0
scikit-learn==1.3.0 